data-pipeline/requirements.txt
psycopg2-binary==2.9.9
mysql-connector-python==9.0.0
python-dotenv==1.0.1
